re-cut base data class from plots and experiments data
