+++ extension/src/plots/data/index.ts
@@ -1,27 +1,10 @@
-import { Disposable } from '@hediet/std/disposable'
-import { Deferred } from '@hediet/std/synchronization'
-import { Event, EventEmitter } from 'vscode'
+import { EventEmitter } from 'vscode'
 import { AvailableCommands, InternalCommands } from '../../commands/internal'
+import { BaseData } from '../../data'
 import { PlotsOutput } from '../../plots/webview/contract'
-import { ProcessManager } from '../../processManager'
 import { sameContents } from '../../util/array'
 
-export class PlotsData {
-  public readonly dispose = Disposable.fn()
-  public readonly onDidUpdate: Event<PlotsOutput>
-
-  protected readonly dvcRoot: string
-  protected readonly processManager: ProcessManager
-
-  private readonly deferred = new Deferred()
-  private readonly initialized = this.deferred.promise
-
-  private readonly internalCommands: InternalCommands
-
-  private readonly updated: EventEmitter<PlotsOutput> = this.dispose.track(
-    new EventEmitter()
-  )
-
+export class PlotsData extends BaseData<PlotsOutput> {
   private revisions?: string[]
 
   constructor(
@@ -29,22 +12,9 @@ export class PlotsData {
     internalCommands: InternalCommands,
     updatesPaused: EventEmitter<boolean>
   ) {
-    this.dvcRoot = dvcRoot
-    this.processManager = this.dispose.track(
-      new ProcessManager(updatesPaused, {
-        name: 'update',
-        process: () => this.update()
-      })
-    )
-
-    this.internalCommands = internalCommands
-    this.onDidUpdate = this.updated.event
+    super(dvcRoot, internalCommands, updatesPaused)
 
     this.initialize()
-  }
-
-  public isReady() {
-    return this.initialized
   }
 
   public setRevisions(...revisions: string[]) {
@@ -56,8 +26,14 @@ export class PlotsData {
     this.managedUpdate()
   }
 
-  public managedUpdate() {
-    return this.processManager.run('update')
+  public async update(): Promise<void> {
+    const data = await this.internalCommands.executeCommand<PlotsOutput>(
+      AvailableCommands.PLOTS_DIFF,
+      this.dvcRoot,
+      ...(this.revisions || [])
+    )
+
+    return this.notifyChanged(data)
   }
 
   private initialize() {
@@ -68,19 +44,5 @@ export class PlotsData {
         this.deferred.resolve()
       })
     )
-  }
-
-  private async update(): Promise<void> {
-    const data = await this.internalCommands.executeCommand<PlotsOutput>(
-      AvailableCommands.PLOTS_DIFF,
-      this.dvcRoot,
-      ...(this.revisions || [])
-    )
-
-    return this.notifyChanged(data)
-  }
-
-  private notifyChanged(data: PlotsOutput) {
-    this.updated.fire(data)
   }
 }
